src/requirements.txt
pandas
openai
unidecode
pyyaml
python-dotenv
tqdm